requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
openai==1.3.8
httpx==0.25.2
python-dotenv==1.0.0
reportlab==4.0.4